Fix 6 code analysis issues. Building this project no longer needs the `findbungs.skip=true` flag  just use `mvn clean install`
